Remove iOS and Android from built platforms before building.
When iOS or Android are included in the .meteor/platforms file, the meteor build command requires the --server option and will fail without it. Given that building for mobile on Bluemix doesn't make sense (that I can think of at the moment), we remove them from .meteor/platforms prior to building.